[Bronze II] Title: 수학은 비대면강의입니다, Time: 244 ms, Memory: 18744 KB -BaekjoonHub

+++ 백준/Bronze/19532. 수학은 비대면강의입니다/수학은 비대면강의입니다.java
@@ -4,7 +4,7 @@ public class Main {
 
 	public static void main(String[] args) {
 		Scanner scn = new Scanner(System.in);
-		long a, b, c, d, e, f;
+		int a, b, c, d, e, f;
 
 		a = scn.nextInt();
 		b = scn.nextInt();
